Class implementing the ImportWireframe interface for .obj files, returns vertices and indices. Code currently untested.

+++ src/se/raddo/raddose3D/ImportWireframe.java
@@ -19,6 +19,6 @@ public interface ImportWireframe {
    * in groups of three indices for each triangle.
    * @return array of indices
    */
-  public double[][] getIndices();
+  public int[][] getIndices();
   
 }
